+++ src/main/java/com/capibaras/bottomline/repository/PayrollRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PayrollRepository extends JpaRepository <Payroll,Integer> {
+public interface PayrollRepository extends JpaRepository <Payroll,Long> {
 }
